Added Winning and Game Over UI to GameScene
Added the UI that allows the player to go back to the main menu if they lose or win the game. It is in the main camera and activates when the game ends. It is just a button that allows the player to Go back to the main menu scene. This is the way the player restarts the game or leaves the game.